% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.dos.numeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Sumatorio de dos números.}
\usage{
suma.dos.numeros(x, y)
}
\arguments{
\item{x}{un dato númerico.}

\item{y}{un dato númerico.}
}
\value{
La suma de \code{x} e \code{y}.
}
\examples{
  suma.dos.numeros(12, 3) 
  suma.dos.numeros(10.8, -13.3)
}
\description{
Sumatorio de dos números.
}
